\docType{data}
\name{Knot_cell_distances.Rda}
\alias{Knot_cell_distances.Rda}
\title{Knot data for process convolution}
\description{
  This simple data set provides distances from each sample unit to each knot (used in process convolutions)
}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{data}

