% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc_STPC.R
\name{mcmc_STPC}
\alias{mcmc_STPC}
\title{function to perform Bayesian analysis of count data using spatio-temporal interactive log Gaussian Cox process model using dynamic process convolution}
\usage{
mcmc_STPC(model, Data, Prior.pars = NULL, Control, Area.adjust = NULL)
}
\arguments{
\item{model}{A formula object specifying the linear predictor for abundance intensity}

\item{Data}{A list holding the following objects:
Adj - Adjacency matrix of ones and zeros describing spatial connectivity (for ICAR modeling)
Grid - SpatialPolygongsDataFrame holding covariate values for all cells one wants to make predictions on (including sampled cells)
Count.data - A data frame with the following columns: (1) "Cell" - entries this column provides the numerical index (in 1:S) for the cell surveyed
      (2) "Time" An integer value providing the time (e.g. day) when sampling occurred, (3) "AreaSurveyed" - provides
      the proportion of the target cell that was surveyed, (4) "Count" Number of target species counted}

\item{Prior.pars}{An optional list giving prior parameters; these include
"beta.tau" precision for Gaussian prior on regression parameters (default 0.1)
"a.eps" alpha parameter for tau_epsilon~Gamma(alpha,beta) (default = 1.0)
"b.eps" beta parameter for tau_epsilon~Gamma(alpha,beta) (default = 0.01)
"a.eta" alpha param for gamma prior precision of spatio-temporal model
"b.eta" beta param for gamma prior precision of spatio-temporal process
"beta0.tau.rw2" prior precision for knot intercepts in rw2 model
"beta1.tau.rw2" prior precision for knot slopes in rw2 model
 If provided, all values for Prior.pars need to be specified}

\item{Control}{A list giving MCMC controls and additional model options
"iter": number of MCMC iterations;
"burnin": number of MCMC burnin iterations;
   "thin": if specified, how many iterations to skip between recorded posterior samples;
"srr.tol": Threshold eigenvalue level for SRR; only eigenvectors with higher eigenvalues than srr.tol are included in SRR formulation (default is 0.5)
"predict" If TRUE (default), calculate posterior predictions across the entire grid
"MH.mu" A vector providing continuous Uniform half-range values for Metropolis-Hastings proposals
"adapt" If true, adapts tuning parameters for nu updates; e.g., one could run a chain with adapt=TRUE, and use adapted MH tuning parameters in a longer analysis (default FALSE)
"fix.tau.epsilon" If TRUE, fixes tau.epsilon to 100}

\item{Area.adjust}{A vector allowing for differences in suitable habitat for each cell.  Can be used for different grid cell sizes or different
proportions of suitable habitat (e.g., 1.0 = 100% of habitat is suitable for the focal species)}
}
\value{
returns a list with the following objecs:
	MCMC: A list object containing posterior samples;
 Accept: A list object indicating the number of proposals that were accepted for parameters updated via Metropolis-Hastings;
 Control: A list object giving MCMC tuning parameters (which are updated if the 'adapt' alorithm is used)
}
\description{
function to perform Bayesian analysis of count data using spatio-temporal interactive log Gaussian Cox process model using dynamic process convolution
}
\author{
Paul B. Conn \email{paul.conn@noaa.gov}
}
\keyword{abundance,}
\keyword{mcmc,}
\keyword{model,}
\keyword{prediction}
\keyword{spatial}
\keyword{spatio-temporal}

