% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_data_Bering.R
\name{sim_data_Bering}
\alias{sim_data_Bering}
\title{function to simulate Bering Sea spotted seal data for spatio-temporal abundance analysis}
\usage{
sim_data_Bering(sim.type)
}
\arguments{
\item{sim.type}{A character string specifying the type of model used for space-time abundance dynamics
("RS2closed" - resource selection on absolute abundance intensity on a closed population; "CPIF" closed population ideal free;
("RS2open" - open population with restricted dis))}
}
\value{
A list object composed of at least two objects, "Data$Grid" is a list vector holding covariate data by day,
        and "Count.data" which holds simulated transect counts
}
\description{
function to simulate Bering Sea spotted seal data for spatio-temporal abundance analysis
}
\author{
Paul B. Conn
}
\keyword{seals}
\keyword{simulation,}
\keyword{spatio-temporal,}
\keyword{spotted}

