% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_data_generic.R
\name{sim_data_generic}
\alias{sim_data_generic}
\title{function to simulate spatio-temporal count data over generic simulated landscapes}
\usage{
sim_data_generic(sim.type = "RS2closed", S, t.steps, n.transects, line.width,
  delta = 0)
}
\arguments{
\item{sim.type}{A character string specifying the type of model used for space-time abundance dynamics
("RS2closed" - resource selection on absolute abundance intensity on a closed population; "CPIF" closed population ideal free;
("RS2open" - open population with restricted dis))}

\item{S}{Number of cells (must be a square number)}

\item{t.steps}{Number of time steps}

\item{n.transects}{Number of transects to simulate at each time step}

\item{line.width}{Proportional of a cell's diameter that is covered by a transect when it surveys a cell}

\item{delta}{Expected proportion increase/decrease for habitat covariate at each time step}
}
\value{
A list object composed of at least two objects, "Data$Grid" is a list vector holding covariate data by day,
        and "Count.data" which holds simulated transect counts
}
\description{
function to simulate spatio-temporal count data over generic simulated landscapes
}
\author{
Paul B. Conn
}
\keyword{simulation}
\keyword{spatio-temporal,}

