% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{sim_effort}
\alias{sim_effort}
\title{function to simulate effort data}
\usage{
sim_effort(S, Data, n.transects = NULL, line.width)
}
\arguments{
\item{S}{Number of cells (must be a square number)}

\item{Data}{A list including at least the following
Grid - A vectored list where each element holds a time-specific SpatialPolygonsDataFrame full of habitat covariates for the area being modeled}

\item{n.transects}{Number of transects to simulate at each time step (set to NULL if effort information is provided instead)}

\item{line.width}{Proportional diameter covered by a transect passing through a cell}
}
\value{
Effort A data.frame holding the following columns:  "Cell", "Time", and "AreaSurveyed"
}
\description{
function to simulate effort data
}
\author{
Paul B. Conn
}
\keyword{simulation}
\keyword{spatio-temporal,}

