% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{sim_Ncounts}
\alias{sim_Ncounts}
\title{function to simulate transect counts given an effort dataset and spatio-temporal lambda values}
\usage{
sim_Ncounts(S, Data, Lambda)
}
\arguments{
\item{S}{Number of cells (must be a square number)}

\item{Data}{A list including at least the following
Grid - A vectored list where each element holds a time-specific SpatialPolygonsDataFrame full of habitat covariates for the area being modeled
Adj - Adjacency matrix for use in ICAR modeling
Effort - Data frame providing records of the amount of area sampled each
   sample unit and time step via columns "Cell", "Time", and "AreaSurveyed".}

\item{Lambda}{A matrix holding spatio-temporal lambda values (sample units on rows, time on columns)}
}
\value{
A listed holding two objects (1) "Count.data" A simulated count dataset, and (2) Simulated true abundance
}
\description{
function to simulate transect counts given an effort dataset and spatio-temporal lambda values
}
\author{
Paul B. Conn
}
\keyword{simulation}
\keyword{spatio-temporal,}

