% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim_funcs.R
\name{sim_RS2}
\alias{sim_RS2}
\title{function to simulate spatio-temporal latent abundance process given an RS2closed formulation}
\usage{
sim_RS2(S, Data, Sim.pars, hab.formula, Area.adjust = NULL)
}
\arguments{
\item{S}{Number of cells (must be a square number)}

\item{Data}{A list including at least the following
Grid - A vectored list where each element holds a time-specific SpatialPolygonsDataFrame full of habitat covariates for the area being modeled
Adj - Adjacency matrix for use in ICAR modeling
Effort - If provided, this data frame provides records of the amount of area sampled each
   sample unit and time step via columns "Cell", "Time", and "AreaSurveyed".  If NULL, transects
   are simulated using the "n.transects" option below
Which.distances - A vector providing the entries of an S by S distance matrix that are nonzero (i.e. those connected in resource selection kernel)
Dist.entries - A vector with specific distance category values (particular to 3 cell radius model in paper)}

\item{Sim.pars}{A list holding parameters that describe evolution of the spatio-temporal process.  Included are
Hab.init - spatial regression parameters for initial time step
Hab.evol - parameters guiding habitat preference associated with the RS process
kern.sd - SD of the redistribution kernel
tau.eta - precision for spatial random effects on initial abundance
tau.epsilon - precision for random noise}

\item{hab.formula}{A formula object holding the regression model formulation}

\item{Area.adjust}{If provided, a vector allowing for differences in suitable habitat for each cell.  Can be used for different grid cell sizes or different
proportions of suitable habitat (e.g., 1.0 = 100% of habitat is suitable for the focal species)}
}
\value{
Lambda  A matrix holding spatio-temporal lambda values
}
\description{
function to simulate spatio-temporal latent abundance process given an RS2closed formulation
}
\author{
Paul B. Conn
}
\keyword{simulation}
\keyword{spatio-temporal,}

